Delete session, course, location OK and Home page Ok

+++ src/main/java/com/utbm/lo54/project/controller/CourseController.java
@@ -3,6 +3,7 @@ package com.utbm.lo54.project.controller;
 import com.utbm.lo54.project.entity.*;
 import com.utbm.lo54.project.service.*;
 import java.util.List;
+import java.util.Objects;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
@@ -12,6 +13,9 @@ import org.springframework.web.servlet.ModelAndView;
 public class CourseController {
     @Autowired
     private CourseService courseService;
+    
+    @Autowired
+    private CourseSessionService courseSessionService;
 
     @RequestMapping("/courses")
     public ModelAndView getAllCourses() {
@@ -46,11 +50,32 @@ public class CourseController {
     	return model;
 
     }
-
+    
+    //Delete course
     @RequestMapping(method = RequestMethod.GET, value = "/courses/deleteCourse/{id}")
     public ModelAndView deleteCourse(@PathVariable Integer id) {
-        courseService.deleteCourse(id);
-        return new ModelAndView("redirect:/courses");
+        Course course = courseService.getCourse(id);
+        Boolean error = true;
+        Boolean test = true;
+        
+        List<CourseSession> sessions = courseSessionService.getAllCoursesSession();
+        for(CourseSession session : sessions) {
+            if(Objects.equals(session.getCourse().getId(), course.getId())){
+                test = false;
+                break;
+            }
+        }
+        if(!test){
+            error = false;
+        }else{
+            error = true;
+            courseService.deleteCourse(id);
+            return new ModelAndView("redirect:/courses");
+        }
+        ModelAndView model = new ModelAndView("body/courses/all");
+        model.addObject("error", error);
+        
+        return model;
     }
     
     //Méthode qui renvoit la vue vers le formulaire d'ajout
