+++ src/main/java/com/utbm/lo54/project/controller/CourseSessionController.java
@@ -2,6 +2,8 @@ package com.utbm.lo54.project.controller;
 
 import com.utbm.lo54.project.entity.*;
 import com.utbm.lo54.project.service.*;
+import java.math.RoundingMode;
+import java.text.DecimalFormat;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -27,7 +29,13 @@ public class CourseSessionController {
     	
     	for(CourseSession session : sessions) {
     		Float percentage = ((float) clientService.CountByCourseSession(session) / session.getMax())*100 ;
-    		session.setBusy(percentage);
+                //Arrondi du pourcentage
+                DecimalFormat df = new DecimalFormat("#.##");
+                df.setRoundingMode(RoundingMode.HALF_UP);
+                 
+                //Float percent = (float)Math.round(percentage * 0);
+                
+    		session.setBusy(df.format(percentage));
     		session.setParticipants(clientService.CountByCourseSession(session));
     	}
     	
@@ -67,8 +75,24 @@ public class CourseSessionController {
 
     @RequestMapping(method = RequestMethod.GET, value = "/session/deleteSession/{id}")
     public ModelAndView deleteCourseSession(@PathVariable Integer id) {
-        courseSessionService.deleteCourseSession(id);
-        return new ModelAndView("redirect:/courseSession");
+        //Récupère la session pour vérifier s'il y a des clients ou pas
+        CourseSession cs = courseSessionService.getCourseSession(id);       
+        List<Client> clients = clientService.findByCourseSession(cs);
+ 
+        Boolean error;
+        
+        if(clients.isEmpty()){
+            error = true;
+            courseSessionService.deleteCourseSession(id);
+            return new ModelAndView("redirect:/courseSession");
+        }else{
+            error = false;
+        }
+        
+        ModelAndView model = new ModelAndView("body/sessions/all");
+        model.addObject("error", error);
+        
+        return model;
     }
     
   //Méthode qui renvoit la vue vers le formulaire d'ajout de la sessioin
