+++ src/main/java/com/utbm/lo54/project/controller/LocationController.java
@@ -3,6 +3,7 @@ package com.utbm.lo54.project.controller;
 import com.utbm.lo54.project.entity.*;
 import com.utbm.lo54.project.service.*;
 import java.util.List;
+import java.util.Objects;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
@@ -11,6 +12,9 @@ import org.springframework.web.servlet.ModelAndView;
 public class LocationController {
     @Autowired
     private LocationService locationService;
+    
+    @Autowired
+    private CourseSessionService courseSessionService;
 
     @RequestMapping("/locations")
     public ModelAndView getAllLocations(){
@@ -41,11 +45,32 @@ public class LocationController {
      	return model;
     	
     }
-
+    //Delete Location
     @RequestMapping(method = RequestMethod.GET, value = "/locations/deleteLocation/{id}")
     public ModelAndView deleteLocation(@PathVariable Integer id) {
-    	locationService.deleteLocation(id);
-    	return new ModelAndView("redirect:/locations");
+    	
+        Location location = locationService.getLocation(id);
+        Boolean error = true;
+        Boolean test = true;
+        
+        List<CourseSession> sessions = courseSessionService.getAllCoursesSession();
+        for(CourseSession session : sessions) {
+            if(Objects.equals(session.getLocation().getId(), location.getId())){
+                test = false;
+                break;
+            }
+        }
+        if(!test){
+            error = false;
+        }else{
+            error = true;
+            locationService.deleteLocation(id);
+            return new ModelAndView("redirect:/locations");
+        }
+        ModelAndView model = new ModelAndView("body/locations/all");
+        model.addObject("error", error);
+        
+        return model;
     }
     //Méthode qui renvoit la vue vers le formulaire d'ajout
     @RequestMapping(method = RequestMethod.GET, value = "/locations/vueAddLocation/")
